refactor for effects
